requirements.txt
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.88.0
gspread==5.9.0
oauth2client==4.1.3
tkinter-tooltip==2.0.1
schedule==1.2.0
python-dotenv==1.0.0
email-validator==2.0.0
pillow==10.0.0
ttkthemes==3.2.2
openpyxl==3.1.2 